BREAKING CHANGES(conventional-changelog): All preset exports factory functions
  * @see: https://github.com/conventional-changelog/conventional-changelog/pull/1045
