+++ packages/conventional-changelog-lmc/src/index.js
@@ -1,9 +1,9 @@
-const parserOpts = require('./parser-opts');
-const writerOpts = require('./writer-opts');
-const recommendedBumpOpts = require('./bump-opts');
+const { createParserOpts } = require('./parserOpts');
+const { createWriterOpts } = require('./writerOpts');
+const { createConventionalRecommendedBumpOpts } = require('./conventionalRecommendedBump');
 
 module.exports = {
-  parserOpts,
-  recommendedBumpOpts,
-  writerOpts,
+  createParserOpts,
+  createWriterOpts,
+  createConventionalRecommendedBumpOpts,
 };
