--- packages/conventional-changelog-lmc/src/parser-opts.js
@@ -1,8 +1,0 @@
-module.exports = {
-  headerPattern: /^(?:Pull request #[0-9]+: )?(?:([a-zA-Z]*-[0-9_]*)(?: ))* ?([\w ]*)(?:\((.*)\))?!?: (.*)$/,
-  breakingHeaderPattern: /^(?:([a-zA-Z]*-[0-9_]*)(?: ))* ?([\w ]*)(?:\((.*)\))?!: (.*)$/,
-  headerCorrespondence: ['body', 'type', 'scope', 'subject'],
-  noteKeywords: ['BREAKING CHANGE', 'BREAKING CHANGES'],
-  revertPattern: /^(?:Revert:)\s"?([\s\S]+?)"?\s*This reverts commit (\w*)\./i,
-  revertCorrespondence: ['header', 'hash'],
-};
